Added Repos: AssuntosRepository, ExemplaresRepository, ObrasRepository

+++ XPTOBusiness/Repositories/IAssuntosRepository.cs
@@ -10,9 +10,9 @@ namespace XPTOBusiness.Repositories
     public interface IAssuntosRepository
     {
         IEnumerable<Assuntos> GetAll();
-        Assuntos GetById(int id);
+        Assuntos GetById(byte id);
         void Add(Assuntos assunto);
         void Update(Assuntos assunto);
-        void Delete(int id);
+        void Delete(byte id);
     }
 }
